installer/requirements_v1.5.1.txt
# WhisperJAV v1.5.1 Requirements
# ================================
#
# This file lists all Python dependencies for WhisperJAV.
# PyTorch is installed separately in post_install_v1.5.1.py
# with CUDA version detection.
#
# Installation: pip install -r requirements_v1.5.1.txt

# ----------------------------------------
# Core Whisper Dependencies
# ----------------------------------------
# Git-based installs from main branches
openai-whisper @ git+https://github.com/openai/whisper@main
stable-ts @ git+https://github.com/meizhong986/stable-ts-fix-setup.git@main

# Whisper variants
faster-whisper>=1.1.0

# ----------------------------------------
# Audio and Signal Processing
# ----------------------------------------
ffmpeg-python         # FFmpeg Python bindings
soundfile             # Audio file I/O
auditok               # Voice Activity Detection (VAD)
librosa               # Audio analysis
pyloudnorm            # Audio normalization

# ----------------------------------------
# GUI Framework (PyWebView)
# ----------------------------------------
pywebview>=4.0        # Modern web-based GUI framework
pythonnet>=3.0        # Required for WebView2 backend on Windows
pywin32; platform_system=="Windows"  # Windows-specific utilities

# ----------------------------------------
# Scientific Computing
# ----------------------------------------
numpy                 # Array operations
scipy                 # Signal processing
matplotlib            # Plotting (for diagnostics)

# ----------------------------------------
# Subtitle Processing
# ----------------------------------------
pysrt                 # SRT file parsing
srt                   # SRT utilities

# ----------------------------------------
# Translation
# ----------------------------------------
PySubtrans>=0.7.0     # AI-powered subtitle translation

# ----------------------------------------
# Utilities
# ----------------------------------------
tqdm                  # Progress bars
aiofiles              # Async file I/O
jsonschema            # Config validation
Pillow                # Image processing
colorama              # Colored console output
requests              # HTTP requests

# ----------------------------------------
# Optional: Performance Optimization
# ----------------------------------------
numba                 # JIT compilation for speedups
